Update: global variables converted from int to byte
byte's consume less memory :)

+++ src/LogicExpressions/PropositionalLogic/Logic/Proposition.java
@@ -89,7 +89,7 @@ public class Proposition {
 
         sentences.add(this.expression.getExpression());
 
-        sentenceCount = operandCount + 1;
+        sentenceCount = (byte) ((int) operandCount + 1);
     }
 
     private void setTruthTable() throws InvalidExpressionException {
